Create separate components for charts
