+++ react-app/src/App.tsx
@@ -2,13 +2,13 @@ import React from "react";
 import { Route } from "react-router-dom";
 import { Home } from "./views/Home";
 import "./App.css";
-import { Charts } from "./views/Charts";
+import { Playground } from "./views/Playground";
 
 function App() {
   return (
     <>
       <Route exact path="/" component={Home} />
-      <Route exact path="/charts" component={Charts} />
+      <Route exact path="/playground" component={Playground} />
     </>
   );
 }
