+++ react-app/src/views/Playground.tsx
@@ -1,0 +1,147 @@
+import React, { FC } from "react";
+import { YieldCurveWidget } from "../components/Widgets/YieldCurveWidget";
+import { HeatMapChart } from "../components/Charts/HeatMapChart";
+
+export const Playground: FC = () => {
+  return (
+    <>
+      <YieldCurveWidget data={lineData} />
+
+      <div style={{ width: 800, height: 600 }} className="mx-auto">
+        <HeatMapChart data={heatMapData} />
+      </div>
+    </>
+  );
+};
+
+const lineData = [
+  {
+    id: "CH",
+    data: [
+      { x: 1, y: -0.00737 },
+      { x: 2, y: -0.00736 },
+      { x: 3, y: -0.00724 },
+      { x: 5, y: -0.00642 },
+      { x: 7, y: -0.00513 },
+      { x: 10, y: -0.00408 },
+      { x: 20, y: -0.00427 },
+      { x: 30, y: 0.00027 },
+      { x: 50, y: 0.00982 },
+      { x: 100, y: 0.01857 },
+    ],
+  },
+
+  {
+    id: "GB",
+    data: [
+      { x: 1, y: 0.00097 },
+      { x: 2, y: 0.00047 },
+      { x: 3, y: 0.00048 },
+      { x: 5, y: 0.001 },
+      { x: 7, y: 0.00153 },
+      { x: 10, y: 0.0023 },
+      { x: 20, y: 0.00314 },
+      { x: 30, y: 0.00295 },
+      { x: 50, y: 0.00217 },
+      { x: 100, y: 0.01801 },
+    ],
+  },
+  {
+    id: "JP",
+    data: [
+      { x: 1, y: -0.00116 },
+      { x: 2, y: -0.00122 },
+      { x: 3, y: -0.00127 },
+      { x: 5, y: -0.00119 },
+      { x: 7, y: -0.00093 },
+      { x: 10, y: -0.0003 },
+      { x: 20, y: 0.00168 },
+      { x: 30, y: 0.00255 },
+      { x: 50, y: 0.01263 },
+      { x: 100, y: 0.02365 },
+    ],
+  },
+  {
+    id: "US",
+    data: [
+      { x: 1, y: 0.00062 },
+      { x: 2, y: 0.00024 },
+      { x: 3, y: 0.00032 },
+      { x: 5, y: 0.00126 },
+      { x: 7, y: 0.00262 },
+      { x: 10, y: 0.00443 },
+      { x: 20, y: 0.00697 },
+      { x: 30, y: 0.00732 },
+      { x: 50, y: 0.00616 },
+      { x: 100, y: 0.02043 },
+    ],
+  },
+];
+
+const heatMapData = [
+  {
+    country: "CH",
+    "1y": -0.737,
+    "2y": -0.736,
+    "3y": -0.724,
+    "5y": -0.642,
+    "7y": -0.513,
+    "10y": -0.408,
+    "20y": -0.427,
+    "30y": 0.027,
+    "50y": 0.982,
+    "100y": 1.857,
+  },
+  {
+    country: "CN",
+    "1y": 1.792,
+    "2y": 1.843,
+    "3y": 1.914,
+    "5y": 2.083,
+    "7y": 2.264,
+    "10y": 2.518,
+    "20y": 3.133,
+    "30y": 3.493,
+    "50y": 3.864,
+    "100y": 4.179,
+  },
+  {
+    country: "GB",
+    "1y": 0.097,
+    "2y": 0.047,
+    "3y": 0.048,
+    "5y": 0.1,
+    "7y": 0.153,
+    "10y": 0.23,
+    "20y": 0.314,
+    "30y": 0.295,
+    "50y": 0.217,
+    "100y": 1.801,
+  },
+  {
+    country: "JP",
+    "1y": -0.116,
+    "2y": -0.122,
+    "3y": -0.127,
+    "5y": -0.119,
+    "7y": -0.093,
+    "10y": -0.03,
+    "20y": 0.168,
+    "30y": 0.255,
+    "50y": 1.263,
+    "100y": 2.365,
+  },
+  {
+    country: "US",
+    "1y": 0.062,
+    "2y": 0.024,
+    "3y": 0.032,
+    "5y": 0.126,
+    "7y": 0.262,
+    "10y": 0.443,
+    "20y": 0.697,
+    "30y": 0.732,
+    "50y": 0.616,
+    "100y": 2.043,
+  },
+];
